+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Improvement in layer priority.
 - Fix bug `animation.playOnce` in Player and Enemy
 - Fix bug `addAction` in `Joystick`
+- Fix bug `seePlayer` in `GameDecoration`
 
 # 1.0.0-rc1